Implement long nicknames/subtitles
Display can be between one and four lines. In JSON mode, use lists of
strings as 'nickname' or 'subtitle' for explicit control of lines, or
keep using strings to have them split at opportune spaces.

In simple mode, it will split long lines at spaces if necessary.

If forced to, it will cut off in the middle of a word. Just insert a
space where you want it to break and it will obey that.